refactor: split secondary infoform commands into infoforms_commands.go
